ADO.NET - SQL Injection. Data Access Layer

+++ practik_a2_2/Product.cs
@@ -6,9 +6,9 @@
         public string? Name { get; set; }
         public string? Type { get; set; }
         public int Quantity { get; set; }
-        public float Cost { get; set; }
+        public int Cost { get; set; }
         public string? Producer { get; set; }
-        public float Price { get; set; }
+        public int Price { get; set; }
         public override string ToString()
         {
             return $"{Name,15}  {Type,15} {Quantity,5} {Cost,5} {Producer,15} {Price,5}";
